Stack has first (half-duplex) only GTS sending/receiving version. Device is able to parse the beacon, understand the GTSs, and send properly. Coordinator is able to create beacon and receive in the whole CFP.
git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@181 88d5ba13-baa4-4915-ae7a-41304cef7dd7

+++ contrib/network/openzb/src/kal/ozb_kal_erika.c
@@ -1,6 +1,6 @@
 #include <kal/ozb_kal_erika.h>
 
-volatile uint16_t ozb_kal_erika_time_counter = 0;
+volatile uint32_t ozb_kal_erika_time_counter = 0;
 
 void ozb_kal_external_timer_action(void) 
 {
